chore: Déplacement des nouvelles fonctions sur les fractions vers la classe FractionX
